Rename MarshallableObject to MarshallableDocument.
Organize types and methods for policy documents so that
types appear at the beginning of the source files.
