+++ pkg/common/policies.go
@@ -5,30 +5,7 @@ import (
 	"time"
 )
 
-// MarshallableObject is an object that can be marshalled and unmarshalled to and from JSON.
-type MarshallableObject interface {
-	Raw() []byte // Returns the Raw JSON this object was unmarshaled from (nil if none).
-}
-
-// PolicyDocument is an interface that is implemented by all objects that are part of the set
-// of "policy objects". A policy object is that one that represents functionality of policies
-// for a domain, such as RPC, RCSR, SPT, SPRT, SP, PSR or Policy.
-type PolicyDocument interface {
-	MarshallableObject
-	Subject() string
-}
-
-type PolicyObjectBase struct {
-	RawJSON    []byte `json:"-"` // omit from JSON (un)marshaling
-	RawSubject string `json:"Subject,omitempty"`
-}
-
-func (o PolicyObjectBase) Raw() []byte     { return o.RawJSON }
-func (o PolicyObjectBase) Subject() string { return o.RawSubject }
-
-// root certificate signing request
-
-// root policy certificate
+// RPC is a Root Policy Certificate.
 type RPC struct {
 	PolicyObjectBase
 	SerialNumber       int                `json:",omitempty"`
@@ -43,6 +20,30 @@ type RPC struct {
 	PRCSignature       []byte             `json:",omitempty"`
 	CASignature        []byte             `json:",omitempty"`
 	SPTs               []SPT              `json:",omitempty"`
+}
+
+// SP is a Signed Policy.
+type SP struct {
+	PolicyObjectBase
+	Policies          DomainPolicy `json:",omitempty"`
+	TimeStamp         time.Time    `json:",omitempty"`
+	CAName            string       `json:",omitempty"`
+	SerialNumber      int          `json:",omitempty"`
+	CASignature       []byte       `json:",omitempty"`
+	RootCertSignature []byte       `json:",omitempty"`
+	SPTs              []SPT        `json:",omitempty"`
+}
+
+// DomainPolicy is a domain policy that specifies what is or not acceptable for a domain.
+type DomainPolicy struct {
+	TrustedCA         []string `json:",omitempty"`
+	AllowedSubdomains []string `json:",omitempty"`
+}
+
+// PCRevocation is for now empty.
+type PCRevocation struct {
+	PolicyObjectBase
+	// TODO(juagargi) define the revocation.
 }
 
 func NewRPC(
@@ -80,30 +81,21 @@ func NewRPC(
 	}
 }
 
-// PCRevocation is for now empty.
-type PCRevocation struct {
-	PolicyObjectBase
-	// TODO(juagargi) define the revocation.
-}
-
-func NewPCRevocation(subject string) *PCRevocation {
-	return &PCRevocation{
-		PolicyObjectBase{
-			RawSubject: subject,
-		},
-	}
-}
-
-// Signed Policy
-type SP struct {
-	PolicyObjectBase
-	Policies          DomainPolicy `json:",omitempty"`
-	TimeStamp         time.Time    `json:",omitempty"`
-	CAName            string       `json:",omitempty"`
-	SerialNumber      int          `json:",omitempty"`
-	CASignature       []byte       `json:",omitempty"`
-	RootCertSignature []byte       `json:",omitempty"`
-	SPTs              []SPT        `json:",omitempty"`
+func (rpc *RPC) Equal(rpc_ *RPC) bool {
+	return true &&
+		rpc.SerialNumber == rpc_.SerialNumber &&
+		rpc.RawSubject == rpc_.RawSubject &&
+		rpc.Version == rpc_.Version &&
+		rpc.PublicKeyAlgorithm == rpc_.PublicKeyAlgorithm &&
+		bytes.Equal(rpc.PublicKey, rpc_.PublicKey) &&
+		rpc.NotBefore.Equal(rpc_.NotBefore) &&
+		rpc.NotAfter.Equal(rpc_.NotAfter) &&
+		rpc.CAName == rpc_.CAName &&
+		rpc.SignatureAlgorithm == rpc_.SignatureAlgorithm &&
+		rpc.TimeStamp.Equal(rpc_.TimeStamp) &&
+		bytes.Equal(rpc.PRCSignature, rpc_.PRCSignature) &&
+		bytes.Equal(rpc.CASignature, rpc_.CASignature) &&
+		equalSPTs(rpc.SPTs, rpc_.SPTs)
 }
 
 func NewSP(
@@ -131,12 +123,6 @@ func NewSP(
 	}
 }
 
-// DomainPolicy is a domain policy.
-type DomainPolicy struct {
-	TrustedCA         []string `json:",omitempty"`
-	AllowedSubdomains []string `json:",omitempty"`
-}
-
 func (s SP) Equal(o SP) bool {
 	return true &&
 		s.TimeStamp.Equal(o.TimeStamp) &&
@@ -149,23 +135,6 @@ func (s SP) Equal(o SP) bool {
 		equalSPTs(s.SPTs, o.SPTs)
 }
 
-func (rpc *RPC) Equal(rpc_ *RPC) bool {
-	return true &&
-		rpc.SerialNumber == rpc_.SerialNumber &&
-		rpc.RawSubject == rpc_.RawSubject &&
-		rpc.Version == rpc_.Version &&
-		rpc.PublicKeyAlgorithm == rpc_.PublicKeyAlgorithm &&
-		bytes.Equal(rpc.PublicKey, rpc_.PublicKey) &&
-		rpc.NotBefore.Equal(rpc_.NotBefore) &&
-		rpc.NotAfter.Equal(rpc_.NotAfter) &&
-		rpc.CAName == rpc_.CAName &&
-		rpc.SignatureAlgorithm == rpc_.SignatureAlgorithm &&
-		rpc.TimeStamp.Equal(rpc_.TimeStamp) &&
-		bytes.Equal(rpc.PRCSignature, rpc_.PRCSignature) &&
-		bytes.Equal(rpc.CASignature, rpc_.CASignature) &&
-		equalSPTs(rpc.SPTs, rpc_.SPTs)
-}
-
 func (s DomainPolicy) Equal(o DomainPolicy) bool {
 	if len(s.TrustedCA) != len(o.TrustedCA) {
 		return false
@@ -176,6 +145,14 @@ func (s DomainPolicy) Equal(o DomainPolicy) bool {
 		}
 	}
 	return true
+}
+
+func NewPCRevocation(subject string) *PCRevocation {
+	return &PCRevocation{
+		PolicyObjectBase{
+			RawSubject: subject,
+		},
+	}
 }
 
 func equalSPTs(a, b []SPT) bool {
